*: use MPMCQueue in UnionBlockInputStream and SharedQueryBlockInputStream (#4418)
close pingcap/tiflash#4419

+++ dbms/src/DataStreams/ParallelInputsProcessor.h
@@ -174,7 +174,7 @@ public:
         return active_threads;
     }
 
-    unsigned int getMaxThreads() const
+    size_t getMaxThreads() const
     {
         return max_threads;
     }
